combined logoutReducer and loginReducer
