+++ src/App.js
@@ -19,12 +19,12 @@ function App() {
   return (
     <div className='App'>
        <NavBar/>
-       <Switch>
+       
          <Route exact path='/' component={Home} />
          <PrivateRoute exact path ='/dashboard' component={Dashboard} />
          <Route exact path='/register' component={Registration} />
          <Route exact path='/login' component={Login} />
-       </Switch>
+       
        <Footer/>
     </div>
   )
