+++ src/reducers/index.js
@@ -3,7 +3,7 @@ import { combineReducers } from 'redux'
 
 //reducers
 import registerReducer from './registerReducer'
-import loginReducer from './loginReducer'
+import userAuthReducer from './userAuthReducer'
 import addBoardReducer from './addBoardReducer'
 import logoutReducer from './logoutReducer'
 import getBoardsReducer from './getBoardsReducer'
@@ -13,7 +13,7 @@ import getArticlesReducer from './getArticlesReducer'
 //root reducer
 const rootReducer = combineReducers({
     registerReducer,
-    loginReducer,
+    userAuthReducer,
     logoutReducer,
     addBoardReducer,
     getBoardsReducer,
