+++ src/reducers/logoutReducer.js
@@ -1,22 +1,22 @@
-import { LOGOUT_START, LOGOUT_SUCCESS, LOGOUT_FAILURE } from '../actions/logoutUser'
 
-const initialState = {
-    inProgress: false,
-    error: ''
-}
 
-function logoutReducer(state = initialState, action) {
-    const { type } = action
-    switch(type) {
-        case LOGOUT_START:
-            return {...state, inProgress: true}
-        case LOGOUT_SUCCESS:
-            return {...state, inProgress: false}
-        case LOGOUT_FAILURE:
-            return {...state, inProgress: false, error: 'Logout Failed'}
-        default:
-            return state
-    }
-}
+// const initialState = {
+//     inProgress: false,
+//     error: ''
+// }
 
-export default logoutReducer
+// function logoutReducer(state = initialState, action) {
+//     const { type } = action
+//     switch(type) {
+//         case LOGOUT_START:
+//             return {...state, inProgress: true}
+//         case LOGOUT_SUCCESS:
+//             return {...state, inProgress: false}
+//         case LOGOUT_FAILURE:
+//             return {...state, inProgress: false, error: 'Logout Failed'}
+//         default:
+//             return state
+//     }
+// }
+
+// export default logoutReducer
